Yaml loader: move tests to its own dir
Simillarly to what was done with the Yaml to mux depedent tests, let's
have tests in the plugin directory itself.
